Damage and DamageDisplay added.
Damage currently stores a float and nothing more. The Player1 and Player2 GameObjects have Damage components attached.

DamageDisplay shows these numbers. It has a damages array, an array of Damage components. If one of this array's values is null, it assumes that player doesn't exist. The DamageGUI array holds the actual Text components to show the numbers. There can be no null values in this array. I've created a GUICanvas GameObject and attached it to the Main Camera GameObject. This GUICanvas object has the Damage Display component attached.